Add files via upload
This document introduces the user manual for the Feitian R301 card reader product, and it works for R301T1 at the moment, and in future, all new Reader from Feitian will support PPDU by default.

In this manual, the feature of obtaining reader information is executed on the Interface Device(IFD) by issuing commands through a Pseudo-Application Protocol Data Unit (PPDU), instead of using the ESCAPE API. On the Windows system, some customers are required to modify the registry to enable the escape channel. To avoid this, we use PPDUs. However, the ESCAPE command has the same functionality. You can still use the ESCAPE command to obtain reader information without any issues. In this manual, we only list the PPDU-related information.